+++ src/mattersim/datasets/utils/converter.py
@@ -1,0 +1,618 @@
+import warnings
+from typing import Optional, Tuple, Literal
+
+import os
+import ase
+import numpy as np
+import torch
+from ase import Atoms
+from pymatgen.optimization.neighbors import find_points_in_spheres
+from pymatgen.core.structure import Structure
+from pymatgen.io.ase import AseAtomsAdaptor
+from torch_geometric.data import Data
+from tqdm import tqdm
+
+from mattersim.datasets.utils.ocp_graph_utils import radius_graph_pbc
+from mattersim.datasets.utils.threebody_indices import compute_threebody as _compute_threebody
+from mattersim.datasets.utils.threebody_indices_torch import (
+    compute_threebody_torch as _compute_threebody_torch,
+)
+
+# Ensure the warning is only shown once
+warnings.filterwarnings("once", category=UserWarning)
+
+"""
+Supported Properties:
+    - "num_nodes"(set by default)  ## int
+    - "num_edges"(set by default)  ## int
+    - "num_atoms"                  ## int
+    - "num_bonds"                  ## int
+    - "atom_attr"                  ## tensor [num_atoms,atom_attr_dim=1]
+    - "atom_pos"                   ## tensor [num_atoms,3]
+    - "edge_length"                ## tensor [num_edges,1]
+    - "edge_vector"                ## tensor [num_edges,3]
+    - "edge_index"                 ## tensor [2,num_edges]
+    - "three_body_indices"         ## tensor [num_three_body,2]
+    - "num_three_body"              ## int
+    - "num_triple_ij"              ## tensor [num_edges,1]
+    - "num_triple_i"               ## tensor [num_atoms,1]
+    - "num_triple_s"               ## tensor [1,1]
+    - "theta_jik"                  ## tensor [num_three_body,1]
+    - "triple_edge_length"         ## tensor [num_three_body,1]
+    - "phi"                        ## tensor [num_three_body,1]
+    - "energy"                     ## float
+    - "forces"                     ## tensor [num_atoms,3]
+    - "stress"                     ## tensor [3,3]
+"""
+
+"""
+Computing various graph based operations (M3GNet)
+"""
+
+
+def compute_threebody_indices(
+    bond_atom_indices: np.array,
+    bond_length: np.array,
+    n_atoms: int,
+    atomic_number: np.array,
+    threebody_cutoff: Optional[float] = None,
+):
+    """
+    Given a graph without threebody indices, add the threebody indices
+    according to a threebody cutoff radius
+    Args:
+        bond_atom_indices: np.array, [n_atoms, 2]
+        bond_length: np.array, [n_atoms]
+        n_atoms: int
+        atomic_number: np.array, [n_atoms]
+        threebody_cutoff: float, threebody cutoff radius
+
+    Returns:
+        triple_bond_indices, n_triple_ij, n_triple_i, n_triple_s
+
+    """
+    n_atoms = np.array(n_atoms).reshape(1)
+    atomic_number = atomic_number.reshape(-1, 1)
+    n_bond = bond_atom_indices.shape[0]
+    if n_bond > 0 and threebody_cutoff is not None:
+        valid_three_body = bond_length <= threebody_cutoff
+        ij_reverse_map = np.where(valid_three_body)[0]
+        original_index = np.arange(n_bond)[valid_three_body]
+        bond_atom_indices = bond_atom_indices[valid_three_body, :]
+    else:
+        ij_reverse_map = None
+        original_index = np.arange(n_bond)
+
+    if bond_atom_indices.shape[0] > 0:
+        bond_indices, n_triple_ij, n_triple_i, n_triple_s = _compute_threebody(
+            np.ascontiguousarray(bond_atom_indices, dtype="int32"),
+            np.array(n_atoms, dtype="int32"),
+        )
+        if ij_reverse_map is not None:
+            n_triple_ij_ = np.zeros(shape=(n_bond,), dtype="int32")
+            n_triple_ij_[ij_reverse_map] = n_triple_ij
+            n_triple_ij = n_triple_ij_
+        bond_indices = original_index[bond_indices]
+        bond_indices = np.array(bond_indices, dtype="int32")
+    else:
+        bond_indices = np.reshape(np.array([], dtype="int32"), [-1, 2])
+        if n_bond == 0:
+            n_triple_ij = np.array([], dtype="int32")
+        else:
+            n_triple_ij = np.array([0] * n_bond, dtype="int32")
+        n_triple_i = np.array([0] * len(atomic_number), dtype="int32")
+        n_triple_s = np.array([0], dtype="int32")
+    return bond_indices, n_triple_ij, n_triple_i, n_triple_s
+
+
+def get_fixed_radius_bonding(
+    structure: ase.Atoms,
+    cutoff: float = 5.0,
+    numerical_tol: float = 1e-8,
+    pbc: bool = True,
+) -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
+    """
+    Get graph representations from structure within cutoff
+    Args:
+        structure (pymatgen Structure or molecule)
+        cutoff (float): cutoff radius
+        numerical_tol (float): numerical tolerance
+
+    Returns:
+        center_indices, neighbor_indices, images, distances
+    """
+    pbc_ = np.array(structure.pbc, dtype=int)
+
+    lattice_matrix = np.ascontiguousarray(structure.cell[:], dtype=float)  # noqa: E501
+
+    cart_coords = np.ascontiguousarray(
+        np.array(structure.positions), dtype=float
+    )  # noqa: E501
+    r = float(cutoff)
+
+    (
+        center_indices,
+        neighbor_indices,
+        images,
+        distances,
+    ) = find_points_in_spheres(  # noqa: E501
+        cart_coords,
+        cart_coords,
+        r=r,
+        pbc=pbc_,
+        lattice=lattice_matrix,
+        tol=numerical_tol,
+    )
+    center_indices = center_indices.astype(np.int64)
+    neighbor_indices = neighbor_indices.astype(np.int64)
+    images = images.astype(np.int64)
+    distances = distances.astype(float)
+    exclude_self = (center_indices != neighbor_indices) | (
+        distances > numerical_tol
+    )  # noqa: E501
+    return (
+        center_indices[exclude_self],
+        neighbor_indices[exclude_self],
+        images[exclude_self],
+        distances[exclude_self],
+    )
+
+
+class GraphConverter:
+    """
+    Convert ase.Atoms to Graph
+    """
+
+    default_properties = ["num_nodes", "num_edges"]
+
+    def __init__(
+        self,
+        model_type: str = "m3gnet",
+        twobody_cutoff: float = 5.0,
+        has_threebody: bool = True,
+        threebody_cutoff: float = 4.0,
+    ):
+        self.model_type = model_type
+        self.twobody_cutoff = twobody_cutoff
+        self.threebody_cutoff = threebody_cutoff
+        self.has_threebody = has_threebody
+
+    def convert(
+        self,
+        atoms: Atoms,
+        energy=None,
+        forces=None,
+        stress=None,
+        pbc=True,
+        **kwargs,
+    ):
+        """
+        Convert the structure into graph
+        Args:
+            pbc: bool, whether to use periodic boundary condition, default True
+        """
+        # normalize the structure
+        if isinstance(atoms, Atoms):
+            pbc_ = np.array(atoms.pbc, dtype=int)
+            if np.all(pbc_ < 0.01) or not pbc:
+                min_x = np.min(atoms.positions[:, 0])
+                min_y = np.min(atoms.positions[:, 1])
+                min_z = np.min(atoms.positions[:, 2])
+                max_x = np.max(atoms.positions[:, 0])
+                max_y = np.max(atoms.positions[:, 1])
+                max_z = np.max(atoms.positions[:, 2])      
+                x_len = (max_x - min_x) + max(
+                    self.twobody_cutoff, self.threebody_cutoff
+                ) * 5
+                y_len = (max_y - min_y) + max(
+                    self.twobody_cutoff, self.threebody_cutoff
+                ) * 5
+                z_len = (max_z - min_z) + max(
+                    self.twobody_cutoff, self.threebody_cutoff
+                ) * 5
+                max_len = max(x_len, y_len, z_len)
+                x_len = y_len = z_len = max_len
+                lattice_matrix = np.eye(3) * max_len
+                pbc_ = np.array([1, 1, 1], dtype=int)
+                warnings.warn(
+                    "No PBC detected, using a large supercell with "
+                    f"size {x_len}x{y_len}x{z_len} Angstrom**3",
+                    UserWarning,
+                )
+               
+                atoms.set_cell(lattice_matrix)
+                atoms.set_pbc(pbc_)
+            else:
+                if np.all(abs(atoms.cell) < 1e-5):
+                    raise ValueError("Cell vectors are too small")
+        else:
+            raise ValueError("structure type not supported")
+
+        scaled_pos = atoms.get_scaled_positions()
+        scaled_pos = np.mod(scaled_pos, 1)
+        atoms.set_scaled_positions(scaled_pos)
+        args = {}
+        if self.model_type == "m3gnet":
+            args["num_atoms"] = len(atoms)
+            args["num_nodes"] = len(atoms)
+            args["atom_attr"] = torch.FloatTensor(
+                atoms.get_atomic_numbers()
+            ).unsqueeze(  # noqa: E501
+                -1
+            )
+            args["atom_pos"] = torch.FloatTensor(atoms.get_positions())
+            args["cell"] = torch.FloatTensor(np.array(atoms.cell)).unsqueeze(0)
+            (
+                sent_index,
+                receive_index,
+                shift_vectors,
+                distances,
+            ) = get_fixed_radius_bonding(atoms, self.twobody_cutoff, pbc=pbc)
+            args["num_bonds"] = len(sent_index)
+            args["edge_index"] = torch.from_numpy(
+                np.array([sent_index, receive_index])
+            )  # noqa: E501
+            args["pbc_offsets"] = torch.FloatTensor(shift_vectors)
+            if self.has_threebody:
+                (
+                    triple_bond_index,
+                    n_triple_ij,
+                    n_triple_i,
+                    n_triple_s,
+                ) = compute_threebody_indices(
+                    bond_atom_indices=args["edge_index"]
+                    .numpy()
+                    .transpose(1, 0),  # noqa: E501
+                    bond_length=distances,
+                    n_atoms=atoms.positions.shape[0],
+                    atomic_number=atoms.get_atomic_numbers(),
+                    threebody_cutoff=self.threebody_cutoff,
+                )
+                args["three_body_indices"] = torch.from_numpy(
+                    triple_bond_index
+                ).to(  # noqa: E501
+                    torch.long
+                )  # [num_three_body,2]
+                args["num_three_body"] = args["three_body_indices"].shape[0]
+                args["num_triple_ij"] = (
+                    torch.from_numpy(n_triple_ij).to(torch.long).unsqueeze(-1)
+                )
+            else:
+                args["three_body_indices"] = None
+                args["num_three_body"] = None
+                args["num_triple_ij"] = None
+            if energy is not None:
+                args["energy"] = torch.FloatTensor([energy])
+            if forces is not None:
+                args["forces"] = torch.FloatTensor(forces)
+            if stress is not None:
+                args["stress"] = torch.FloatTensor(stress).unsqueeze(0)
+            return Data(**args)
+
+        elif self.model_type == "graphormer":
+            raise NotImplementedError
+        else:
+            raise NotImplementedError(
+                "model type {} not implemented".format(self.model_type)
+            )
+
+
+def compute_threebody_indices_torch(
+    edge_indices: torch.Tensor,
+    distances: torch.Tensor,
+    num_atoms: torch.Tensor,
+    threebody_cutoff: float = 4.0,
+):
+    """
+    Given a graph without threebody indices, add the threebody indices
+    according to a threebody cutoff radius
+    Args:
+        edge_indices: torch.Tensor, (2, n_edges)
+        distances: torch.Tensor, (n_edges,)
+        num_atoms: torch.Tensor, (n_structures,)
+        threebody_cutoff: float, threebody cutoff radius, default 4.0
+    Returns:
+        triple_bond_indices, n_triple_ij, n_triple_i, n_triple_s
+    """
+    num_edges = edge_indices.shape[1]
+    total_num_atoms = num_atoms.sum().item()
+    if num_edges > 0 and threebody_cutoff is not None:
+        valid_three_body = distances <= threebody_cutoff
+        ij_reverse_map = torch.where(valid_three_body)[0]
+        original_index = torch.arange(num_edges, device=edge_indices.device)[valid_three_body]
+        valid_edge_indices = edge_indices[:, valid_three_body].transpose(0, 1)
+    else:
+        ij_reverse_map = None
+        original_index = torch.arange(num_edges, device=edge_indices.device)
+
+    if num_edges > 0 and valid_edge_indices.shape[0] > 0:
+        (
+            angle_indices,
+            num_angles_per_edge,
+            num_edges_per_atom,
+            num_angles_per_structure,
+        ) = _compute_threebody_torch(
+            valid_edge_indices,
+            total_num_atoms,
+        )
+        if ij_reverse_map is not None:
+            num_angles_per_edge_ = torch.zeros(
+                (num_edges,), dtype=torch.long, device=edge_indices.device
+            )
+            num_angles_per_edge_[ij_reverse_map] = num_angles_per_edge
+            num_angles_per_edge = num_angles_per_edge_
+        angle_indices = original_index[angle_indices]
+    else:
+        # No valid edges
+        angle_indices = torch.zeros((0, 2), dtype=torch.long, device=edge_indices.device)
+        if num_edges == 0:
+            num_angles_per_edge = torch.zeros((0,), dtype=torch.int32, device=edge_indices.device)
+        else:
+            num_angles_per_edge = torch.zeros(
+                (num_edges,), dtype=torch.int32, device=edge_indices.device
+            )
+        num_edges_per_atom = torch.zeros(
+            (total_num_atoms,), dtype=torch.int32, device=edge_indices.device
+        )
+        num_angles_per_structure = torch.zeros(
+            (num_atoms.shape[0],), dtype=torch.int32, device=edge_indices.device
+        )
+    return angle_indices, num_angles_per_edge, num_edges_per_atom, num_angles_per_structure
+
+
+
+class BatchGraphConverter:
+    """
+    Convert batch of materials to graphs, with GPU support
+    
+    Args:
+        model_type: Literal["m3gnet"], "m3gnet"
+        twobody_cutoff: float cutoff for two body interactions, default 5.0 angstrom
+        has_threebody: bool, whether to compute three body interactions, only for m3gnet, default True
+        threebody_cutoff: float cutoff for three body interactions, default 4.0 angstrom
+    """
+
+    default_properties = ["num_nodes", "num_edges"]
+
+    def __init__(
+        self,
+        model_type: Literal["m3gnet"] = "m3gnet",
+        twobody_cutoff: float = 5.0,
+        has_threebody: bool = True,
+        threebody_cutoff: float = 4.0,
+        max_num_neighbors_threshold: int = int(1e6),
+        device: str | torch.device | None = None,
+    ):
+        self.model_type = model_type
+        self.twobody_cutoff = twobody_cutoff
+        self.threebody_cutoff = threebody_cutoff
+        self.has_threebody = has_threebody
+        self.max_num_neighbors_threshold = max_num_neighbors_threshold
+        if device is None:
+            self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+        elif isinstance(device, str):
+            self.device = torch.device(device)
+        else:
+            self.device = device
+        if model_type != "m3gnet":
+            raise NotImplementedError(
+                "BatchGraphConverter only supports m3gnet model type"
+            )
+
+    def convert(
+        self,
+        atoms_list: Atoms | Structure | list[Atoms | Structure],
+        *,
+        energy: list[float] | None = None,
+        forces: list[np.ndarray] | None = None,
+        stresses: list[np.ndarray] | None = None,
+        max_natoms_per_batch: int = 4096,
+        # pbc: bool | list[bool] = True,
+    ):
+        """
+        Convert the structure into graph
+        Args:
+            atoms_list: Atoms | Structure or list of Atoms | Structure
+            energy: list of float, energies for each structure
+            forces: list of np.ndarray, forces for each structure
+            stresses: list of np.ndarray, stress for each structure
+            max_natoms_per_batch: int, maximum number of atoms per batch
+        """
+        if isinstance(atoms_list, (Atoms, Structure)):
+            atoms_list = [atoms_list]
+        if isinstance(atoms_list[0], Structure):
+            atoms_list = [AseAtomsAdaptor.get_atoms(structure) for structure in atoms_list]
+
+        graphs = []
+        pointer = 0
+        pbar = tqdm(
+            total=len(atoms_list),
+            desc="Converting to graphs",
+            disable=os.environ.get("DEBUG") not in ["1", "DEBUG"],
+        )
+        while pointer < len(atoms_list):
+            natoms = torch.zeros((0,), dtype=torch.long, device=self.device)
+            pos = torch.zeros((0, 3), dtype=torch.float32, device=self.device)
+            cell = torch.zeros((0, 3, 3), dtype=torch.float32, device=self.device)
+            atomic_numbers = torch.zeros((0,), dtype=torch.long, device=self.device)
+            pbc = torch.zeros((0, 3), dtype=torch.bool, device=self.device)
+            natoms_cumsum = 0
+            num_graphs = 0
+            while (
+                pointer < len(atoms_list)
+                and natoms_cumsum + len(atoms_list[pointer]) <= max_natoms_per_batch
+            ):
+                atoms = atoms_list[pointer].copy()
+                atoms.wrap()
+                natoms = torch.cat(
+                    (natoms, torch.tensor([len(atoms)], dtype=torch.long, device=self.device)),
+                    dim=0,
+                )
+                pos = torch.cat(
+                    (
+                        pos,
+                        torch.tensor(
+                            atoms.get_positions(), dtype=torch.float32, device=self.device
+                        ),
+                    ),
+                    dim=0,
+                )
+                cell = torch.cat(
+                    (
+                        cell,
+                        torch.tensor(
+                            np.array(atoms.cell), dtype=torch.float32, device=self.device
+                        ).unsqueeze(0),
+                    ),
+                    dim=0,
+                )
+                atomic_numbers = torch.cat(
+                    (
+                        atomic_numbers,
+                        torch.tensor(
+                            atoms.get_atomic_numbers(), dtype=torch.long, device=self.device
+                        ),
+                    ),
+                    dim=0,
+                )
+                pbc = torch.cat(
+                    (
+                        pbc,
+                        torch.tensor(
+                            np.array([atoms.pbc], dtype=bool), dtype=torch.bool, device=self.device
+                        ),
+                    ),
+                    dim=0,
+                )
+                natoms_cumsum += len(atoms)
+                pointer += 1
+                num_graphs += 1
+                pbar.update(1)
+
+            if num_graphs == 0:
+                # Handle the case where a single graph exceeds max_natoms_per_batch anyway
+                num_graphs = 1
+                atoms = atoms_list[pointer].copy()
+                atoms.wrap()
+                natoms = torch.cat(
+                    (natoms, torch.tensor([len(atoms)], dtype=torch.long, device=self.device)),
+                    dim=0,
+                )
+                pos = torch.cat(
+                    (
+                        pos,
+                        torch.tensor(
+                            atoms.get_positions(), dtype=torch.float32, device=self.device
+                        ),
+                    ),
+                    dim=0,
+                )
+                cell = torch.cat(
+                    (
+                        cell,
+                        torch.tensor(
+                            np.array(atoms.cell), dtype=torch.float32, device=self.device
+                        ).unsqueeze(0),
+                    ),
+                    dim=0,
+                )
+                atomic_numbers = torch.cat(
+                    (
+                        atomic_numbers,
+                        torch.tensor(
+                            atoms.get_atomic_numbers(), dtype=torch.long, device=self.device
+                        ),
+                    ),
+                    dim=0,
+                )
+                pbc = torch.cat(
+                    (
+                        pbc,
+                        torch.tensor(
+                            np.array(atoms.pbc, dtype=bool), dtype=torch.bool, device=self.device
+                        ),
+                    ),
+                    dim=0,
+                )
+                natoms_cumsum += len(atoms)
+                pointer += 1
+                pbar.update(1)
+
+            edge_indices, offsets, num_edges, _, distances = radius_graph_pbc(
+                pos=pos,
+                pbc=pbc,
+                cell=cell,
+                natoms=natoms,
+                radius=self.twobody_cutoff,
+                max_num_neighbors_threshold=self.max_num_neighbors_threshold,
+            )
+            edge_indices = torch.cat(
+                (edge_indices[1].unsqueeze(0), edge_indices[0].unsqueeze(0)), dim=0
+            )
+            num_edges = num_edges.to(torch.long)
+
+            (
+                triple_bond_indices,
+                n_triple_ij,
+                n_triple_i,
+                n_triple_s,
+            ) = compute_threebody_indices_torch(
+                edge_indices=edge_indices,
+                distances=distances,
+                num_atoms=natoms,
+                threebody_cutoff=self.threebody_cutoff,
+            )
+
+            # split the converted graphs into list of Data objects for each structure
+            start_edge = 0
+            start_atom = 0
+            for i in range(num_graphs):
+                graph = {}
+                n_atoms_i = natoms[i].item()
+                n_edges_i = num_edges[i].item()
+                graph["num_atoms"] = n_atoms_i
+                graph["num_nodes"] = n_atoms_i
+                graph["atom_attr"] = (
+                    atomic_numbers[start_atom : start_atom + n_atoms_i]
+                    .unsqueeze(-1)
+                    .to(torch.float32)
+                )
+                graph["atom_pos"] = pos[start_atom : start_atom + n_atoms_i]
+                graph["cell"] = cell[i].unsqueeze(0)
+                graph["num_bonds"] = n_edges_i
+                graph["edge_index"] = (
+                    edge_indices[:, start_edge : (start_edge + n_edges_i)] - start_atom
+                )
+                graph["distances"] = distances[start_edge : start_edge + n_edges_i]
+                graph["pbc_offsets"] = offsets[start_edge : start_edge + n_edges_i]
+                if self.has_threebody:
+                    n_triple_ij_i = n_triple_ij[start_edge : start_edge + n_edges_i]
+                    mask = (triple_bond_indices[:, 0] >= start_edge) & (
+                        triple_bond_indices[:, 0] < start_edge + n_edges_i
+                    )
+                    triple_bond_indices_i = triple_bond_indices[mask, :] - start_edge
+                    graph["three_body_indices"] = triple_bond_indices_i
+                    graph["num_three_body"] = triple_bond_indices_i.shape[0]
+                    graph["num_triple_ij"] = n_triple_ij_i.unsqueeze(-1)
+                else:
+                    graph["three_body_indices"] = None
+                    graph["num_three_body"] = None
+                    graph["num_triple_ij"] = None
+
+                if energy is not None and energy[pointer - num_graphs + i] is not None:
+                    graph["energy"] = torch.tensor(
+                        [energy[pointer - num_graphs + i]], dtype=torch.float32, device=self.device
+                    )
+                if forces is not None and forces[pointer - num_graphs + i] is not None:
+                    graph["forces"] = torch.tensor(
+                        forces[pointer - num_graphs + i], dtype=torch.float32, device=self.device
+                    )
+                if stresses is not None and stresses[pointer - num_graphs + i] is not None:
+                    graph["stress"] = torch.tensor(
+                        stresses[pointer - num_graphs + i], dtype=torch.float32, device=self.device
+                    ).unsqueeze(0)
+                graphs.append(Data(**graph).to(self.device))
+                start_edge += n_edges_i
+                start_atom += n_atoms_i
+
+        return graphs
